Move AppMessage class out of context
This didn't really make too much sense being defined alongside the
MessageContext, so I moved this into a separate directory/file. Now
I'm not sure if the standalone context directory still makes sense,
but since those will be imported into multiple components I think it's
better to have them in a consolidated place.

I also changed the `message` property of the `AppMessage` to be `text`
instead, to avoid duplicating the word message
